various flow improvements/changes to dsym functionality - mainly to how defaults work etc and path processing
